% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_converter.R
\name{translate_knr}
\alias{translate_knr}
\title{Translate municipal numbers from one time point to another}
\usage{
translate_knr(knr, from_date, to_date, show_warnings = TRUE)
}
\arguments{
\item{knr}{Character. The four-digits municipal number.}

\item{from_date}{The date from which the knr is.}

\item{to_date}{The date in which to get the knr for.}
}
\value{
A character vector with the municipal numbers used on to_date.
}
\description{
Translate municipal numbers from one time point to another
}
